adding traits
