+++ src/Omnipay/Realex/Message/GenerateAuthorizeTokenResponse.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace Omnipay\Realex\Message;
+
+use Omnipay\Common\Message\AbstractRequest;
+
+class GenerateTokenResponse extends AbstractRequest
+{
+    private $data;
+
+    public function __construct($data)
+    {
+        $this->data = $data;
+    }
+
+    public function getData()
+    {
+        return $this->data;
+    }
+
+    public function sendData($data)
+    {
+
+    }
+}
